Python Project
This project brings up a graphical user interface that the user can easily go through, in this case it is an Intel workshop instructor, and enter 3 simple inputs. With these inputs, the program creates a new folder with a template Excel macro file in it, then opens Chrome automatically and creates a survey on Qualtrics and retrieves the anonymous link for the survey, and places it into the Excel macro file.